Tighten chat sidebar and artifact contracts
